backend/requirements.txt
flask
pymongo
authlib
flask-cors
flask-mail
pyjwt
twilio
werkzeug
dotenv
torch
pillow
numpy
opencv-python
opencv-contrib-python
torchvision
